+++ packages/react-learn-card/src/components/LearnCardCreditCardFrontFace/LearnCardCreditCardFrontFace.stories.tsx
@@ -24,4 +24,5 @@ LearnCardCreditCardFrontFaceTest.args = {
     actionButtonText: 'Open Card',
     onClick: () => {},
     className: '',
+    children: <></>,
 };