+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/snap-chapi-example
 
+## 1.0.3
+
+### Patch Changes
+
+-   Updated dependencies [[`426702f`](https://github.com/learningeconomy/LearnCard/commit/426702f50b8790a8eeb68908331a79c79043f4f5)]:
+    -   @learncard/react@2.3.33
+
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies [[`4e5cc2f`](https://github.com/learningeconomy/LearnCard/commit/4e5cc2fe935f99ca663bc9b7d75db2a86c0b7b23)]:
+    -   @learncard/react@2.3.32
+
 ## 1.0.1
 
 ### Patch Changes